+++ chromium/chrome/installer/util/app_command.cc
@@ -4,7 +4,10 @@
 
 #include "chrome/installer/util/app_command.h"
 
+#include <stddef.h>
+
 #include "base/logging.h"
+#include "base/macros.h"
 #include "base/win/registry.h"
 #include "chrome/installer/util/google_update_constants.h"
 #include "chrome/installer/util/work_item_list.h"
@@ -57,7 +60,7 @@ bool AppCommand::Initialize(const base::win::RegKey& key) {
 
   command_line_.swap(cmd_line);
 
-  for (int i = 0; i < arraysize(kNameBoolVars); ++i) {
+  for (size_t i = 0; i < arraysize(kNameBoolVars); ++i) {
     DWORD value = 0;  // Set default to false.
     // Note: ReadValueDW only modifies out param on success.
     key.ReadValueDW(kNameBoolVars[i].name, &value);
@@ -83,7 +86,7 @@ void AppCommand::AddWorkItems(HKEY predefined_root,
                                     true)
       ->set_log_message("setting AppCommand CommandLine registry value");
 
-  for (int i = 0; i < arraysize(kNameBoolVars); ++i) {
+  for (size_t i = 0; i < arraysize(kNameBoolVars); ++i) {
     const wchar_t* var_name = kNameBoolVars[i].name;
     bool var_data = this->*(kNameBoolVars[i].data);
