+++ chromium/chrome/installer/util/installation_validator.cc
@@ -6,11 +6,14 @@
 
 #include "chrome/installer/util/installation_validator.h"
 
+#include <stddef.h>
+
 #include <algorithm>
 #include <set>
 #include <string>
 
 #include "base/logging.h"
+#include "base/macros.h"
 #include "base/strings/utf_string_conversions.h"
 #include "base/version.h"
 #include "chrome/common/chrome_switches.h"
@@ -152,7 +155,7 @@ void InstallationValidator::ValidateAppCommandFlags(
          app_cmd.is_run_as_user(),
          "be marked to run as user"},
   };
-  for (int i = 0; i < arraysize(check_list); ++i) {
+  for (size_t i = 0; i < arraysize(check_list); ++i) {
     bool expected = flags_exp.find(check_list[i].exp_key) != flags_exp.end();
     if (check_list[i].val != expected) {
       *is_valid = false;
