+++ chromium/chrome/installer/util/self_cleaning_temp_dir_unittest.cc
@@ -3,10 +3,12 @@
 // found in the LICENSE file.
 
 #include <windows.h>
+#include <stdint.h>
 #include <wincrypt.h>
 
 #include "base/files/file_util.h"
 #include "base/files/scoped_temp_dir.h"
+#include "base/macros.h"
 #include "base/strings/string_number_conversions.h"
 #include "chrome/installer/util/self_cleaning_temp_dir.h"
 #include "testing/gtest/include/gtest/gtest.h"
@@ -16,7 +18,7 @@ namespace {
 // Returns a string of 8 characters consisting of the letter 'R' followed by
 // seven random hex digits.
 std::string GetRandomFilename() {
-  uint8 data[4];
+  uint8_t data[4];
   HCRYPTPROV crypt_ctx = NULL;
 
   // Get four bytes of randomness.  Use CAPI rather than the CRT since I've
@@ -28,7 +30,7 @@ std::string GetRandomFilename() {
 
   // Hexify the value.
   std::string result(base::HexEncode(&data[0], arraysize(data)));
-  EXPECT_EQ(8, result.size());
+  EXPECT_EQ(8u, result.size());
 
   // Replace the first digit with the letter 'R' (for "random", get it?).
   result[0] = 'R';
@@ -159,7 +161,7 @@ TEST_F(SelfCleaningTempDirTest, LeaveUsedOnDestroy) {
     EXPECT_EQ(parent_temp_dir.Append(L"Three"), temp_dir.path());
     EXPECT_TRUE(base::DirectoryExists(temp_dir.path()));
     // Drop a file somewhere.
-    EXPECT_EQ(arraysize(kHiHon) - 1,
+    EXPECT_EQ(static_cast<int>(arraysize(kHiHon) - 1),
               base::WriteFile(parent_temp_dir.AppendASCII(GetRandomFilename()),
                               kHiHon, arraysize(kHiHon) - 1));
   }
