+++ chromium/content/browser/system_message_window_win.cc
@@ -5,8 +5,10 @@
 #include "content/browser/system_message_window_win.h"
 
 #include <dbt.h>
+#include <stddef.h>
 
 #include "base/logging.h"
+#include "base/macros.h"
 #include "base/system_monitor/system_monitor.h"
 #include "base/win/wrapped_window_proc.h"
 #include "media/audio/win/core_audio_util_win.h"
@@ -45,7 +47,7 @@ class SystemMessageWindowWin::DeviceNotifications {
     filter.dbcc_size = sizeof(filter);
     filter.dbcc_devicetype = DBT_DEVTYP_DEVICEINTERFACE;
     bool core_audio_support = media::CoreAudioUtil::IsSupported();
-    for (int i = 0; i < arraysize(kDeviceCategoryMap); ++i) {
+    for (size_t i = 0; i < arraysize(kDeviceCategoryMap); ++i) {
       // If CoreAudio is supported, AudioDeviceListenerWin will
       // take care of monitoring audio devices.
       if (core_audio_support &&
@@ -63,7 +65,7 @@ class SystemMessageWindowWin::DeviceNotifications {
   }
 
   void Unregister() {
-    for (int i = 0; i < arraysize(notifications_); ++i) {
+    for (size_t i = 0; i < arraysize(notifications_); ++i) {
       if (notifications_[i]) {
         UnregisterDeviceNotification(notifications_[i]);
         notifications_[i] = NULL;
@@ -122,7 +124,7 @@ LRESULT SystemMessageWindowWin::OnDeviceChange(UINT event_type, LPARAM data) {
           reinterpret_cast<DEV_BROADCAST_DEVICEINTERFACE*>(data);
       if (device_interface->dbcc_devicetype != DBT_DEVTYP_DEVICEINTERFACE)
         return TRUE;
-      for (int i = 0; i < arraysize(kDeviceCategoryMap); ++i) {
+      for (size_t i = 0; i < arraysize(kDeviceCategoryMap); ++i) {
         if (kDeviceCategoryMap[i].device_category ==
             device_interface->dbcc_classguid) {
           device_type = kDeviceCategoryMap[i].device_type;
